Secure-Document-Platform/report-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-jose
reportlab